Configure ssh-add in geode-main, not .tmux.conf
Don't add SSH_AUTH_SOCK to tmux update-environment.  It appears to be
part of the default value for that option now.  It's probably not what I
want anyway, because tmux won't pass the variable to new sessions unless
it was already set in the client environment.  This commit works around
that by setting the variable in the client environment explicitly before
connecting to tmux.